Move `SettingsListener` into `ServiceEndpoint`
