+++ android/src/main/kotlin/net/mullvad/mullvadvpn/service/AccountCache.kt
@@ -2,6 +2,7 @@ package net.mullvad.mullvadvpn.service
 
 import kotlinx.coroutines.delay
 import net.mullvad.mullvadvpn.model.GetAccountDataResult
+import net.mullvad.mullvadvpn.service.endpoint.SettingsListener
 import net.mullvad.mullvadvpn.util.ExponentialBackoff
 import net.mullvad.mullvadvpn.util.JobTracker
 import net.mullvad.talpid.util.EventNotifier
