+++ android/src/main/kotlin/net/mullvad/mullvadvpn/service/ServiceInstance.kt
@@ -1,6 +1,7 @@
 package net.mullvad.mullvadvpn.service
 
 import android.os.Messenger
+import net.mullvad.mullvadvpn.service.endpoint.SettingsListener
 import net.mullvad.talpid.ConnectivityListener
 
 class ServiceInstance(
@@ -22,6 +23,5 @@ class ServiceInstance(
         customDns.onDestroy()
         keyStatusListener.onDestroy()
         locationInfoCache.onDestroy()
-        settingsListener.onDestroy()
     }
 }
